feat: implement proper axis reduction and optimize array ops Fix the half-assed nc_sum() implementation to actually handle arbitrary axis reduction instead of just the trivial case. Replace the crappy generic value assignment with type-specific tight loops in nc_arange() to stop wasting CPU cycles on unnecessary function calls.
Move _dtype_size() where it belongs and unfuck the header organization.
Stop embedding macros where they don't belong - put that crap in macros.h
like a sane person would.

Added proper benchmark example that actually tests something useful.
The old one was pure garbage that even an intern would be ashamed of.

+++ benchmarks/py_bm.py
@@ -3,7 +3,7 @@ import numpy as np
 
 start = time.perf_counter()
 
-test = np.arange(1, 100_000_001, dtype=np.int32)
+test = np.arange(1, 100_000_001)
 test = test.reshape((100, 100, 100, 100))
 sall = np.sum(test, 3)
